[KOGITO-3039] - Incorporate review comments
