+++ pkg/controller/kogitoinfra/kogitoinfra_service.go
@@ -17,28 +17,10 @@ package kogitoinfra
 import (
 	"fmt"
 	"github.com/kiegroup/kogito-cloud-operator/pkg/apis/app/v1alpha1"
-	"github.com/kiegroup/kogito-cloud-operator/pkg/client"
-	"github.com/kiegroup/kogito-cloud-operator/pkg/client/kubernetes"
 	"github.com/kiegroup/kogito-cloud-operator/pkg/controller/kogitoinfra/infinispan"
 	"github.com/kiegroup/kogito-cloud-operator/pkg/controller/kogitoinfra/kafka"
 	"github.com/kiegroup/kogito-cloud-operator/pkg/controller/kogitoinfra/keycloak"
-	"k8s.io/apimachinery/pkg/types"
 )
-
-// FetchKogitoInfraInstance load infra instance
-func FetchKogitoInfraInstance(client *client.Client, name string, namespace string) (*v1alpha1.KogitoInfra, error) {
-	log.Debugf("going to fetch deployed kogito infra instance %s", name)
-	instance := &v1alpha1.KogitoInfra{}
-	if exists, resultErr := kubernetes.ResourceC(client).FetchWithKey(types.NamespacedName{Name: name, Namespace: namespace}, instance); resultErr != nil {
-		log.Errorf("Error occurs while fetching deployed kogito infra instance %s", name)
-		return nil, resultErr
-	} else if !exists {
-		return nil, fmt.Errorf("kogito Infra resource with name %s not found in namespace %s", name, namespace)
-	} else {
-		log.Debugf("Successfully fetch deployed kogito infra reference %s", name)
-		return instance, nil
-	}
-}
 
 // GetKogitoInfraResource identify and return request kogito infra resource on bases of information provided in kogitoInfra value
 func GetKogitoInfraResource(instance *v1alpha1.KogitoInfra) (InfraResource, error) {
